Refactor match_options_widget.py to external file.
